Videos/2026-03-09_DenchClaw_Local_AI_Agents_for_CRM__Outreach/requirements.txt
pandas>=1.3.0
transformers>=4.21.0
torch>=1.12.0
pyyaml>=6.0
